// src/frontend/src/common/graphql/types.ts
export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type Chat = {
  __typename: "Chat",
  userId?: string | null,
  sessionId: string,
  human?: string | null,
  assistant?: string | null,
  trace?: string | null,
  traceMetadata?: string | null,
  expiration?: number | null,
  id: string,
  createdAt: string,
  updatedAt: string,
  sessionChatsId?: string | null,
};
// ...
export type Session = {
  __typename: "Session",
  userId?: string | null,
  chats?: ModelChatConnection | null,
  expiration?: number | null,
  id: string,
  createdAt: string,
  updatedAt: string,
};

export type ModelChatConnection = {
  __typename: "ModelChatConnection",
  items:  Array<Chat | null >,
  nextToken?: string | null,
};

export type UpdateSessionInput = {
  userId?: string | null,
  expiration?: number | null,
  id: string,
};

export type DeleteSessionInput = {
  id: string,
};

export type ModelChatFilterInput = {
  userId?: ModelIDInput | null,
  sessionId?: ModelStringInput | null,
  human?: ModelStringInput | null,
  assistant?: ModelStringInput | null,
  trace?: ModelStringInput | null,
  traceMetadata?: ModelStringInput | null,
  expiration?: ModelIntInput | null,
  id?: ModelIDInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  and?: Array< ModelChatFilterInput | null > | null,
  or?: Array< ModelChatFilterInput | null > | null,
  not?: ModelChatFilterInput | null,
  sessionChatsId?: ModelIDInput | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}


export type ModelSessionFilterInput = {
  userId?: ModelIDInput | null,
  expiration?: ModelIntInput | null,
  id?: ModelIDInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  and?: Array< ModelSessionFilterInput | null > | null,
  or?: Array< ModelSessionFilterInput | null > | null,
  not?: ModelSessionFilterInput | null,
};

export type ModelSessionConnection = {
  __typename: "ModelSessionConnection",
  items:  Array<Session | null >,
  nextToken?: string | null,
};
// ...
export type GetChatQueryVariables = {
  id: string,
};

export type GetChatQuery = {
  getChat?:  {
    __typename: "Chat",
    userId?: string | null,
    sessionId: string,
    human?: string | null,
    assistant?: string | null,
    trace?: string | null,
    traceMetadata?: string | null,
    expiration?: number | null,
    id: string,
    createdAt: string,
    updatedAt: string,
    sessionChatsId?: string | null,
  } | null,
};

export type ListChatsQueryVariables = {
  filter?: ModelChatFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListChatsQuery = {
  listChats?:  {
    __typename: "ModelChatConnection",
    items:  Array< {
      __typename: "Chat",
      userId?: string | null,
      sessionId: string,
      human?: string | null,
      assistant?: string | null,
      trace?: string | null,
      traceMetadata?: string | null,
      expiration?: number | null,
      id: string,
      createdAt: string,
      updatedAt: string,
      sessionChatsId?: string | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ChatsBySessionIdQueryVariables = {
  sessionId: string,
  sortDirection?: ModelSortDirection | null,
  filter?: ModelChatFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ChatsBySessionIdQuery = {
  chatsBySessionId?:  {
    __typename: "ModelChatConnection",
    items:  Array< {
      __typename: "Chat",
      userId?: string | null,
      sessionId: string,
      human?: string | null,
      assistant?: string | null,
      trace?: string | null,
      traceMetadata?: string | null,
      expiration?: number | null,
      id: string,
      createdAt: string,
      updatedAt: string,
      sessionChatsId?: string | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type GetSessionQueryVariables = {
  id: string,
};

export type GetSessionQuery = {
  getSession?:  {
    __typename: "Session",
    userId?: string | null,
    chats?:  {
      __typename: "ModelChatConnection",
      nextToken?: string | null,
    } | null,
    expiration?: number | null,
    id: string,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type ListSessionsQueryVariables = {
  filter?: ModelSessionFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListSessionsQuery = {
  listSessions?:  {
    __typename: "ModelSessionConnection",
    items:  Array< {
      __typename: "Session",
      userId?: string | null,
      expiration?: number | null,
      id: string,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};